+++ variants/vite_tailwind/config/routes.rb
@@ -1,0 +1,9 @@
+Rails.application.routes.draw do
+  draw :pages
+  draw :landing
+  draw :errors
+  draw :dev
+  draw :support
+
+  root "landing#home"
+end