+++ variants/vite_tailwind/config/routes/dev.rb
@@ -1,0 +1,7 @@
+if Rails.env.development?
+  namespace :dev do
+    mount LetterOpenerWeb::Engine, at: "/letter_opener"
+    mount MissionControl::Jobs::Engine, at: "/jobs"
+    mount Lookbook::Engine, at: "/lookbook"
+  end
+end